+++ fuc/cli/bam_depth.py
@@ -5,11 +5,15 @@ from .. import api
 import pysam
 
 description = f"""
-This command will compute read depth from input SAM/BAM/CRAM files.
+###############################################
+# Compute read depth from SAM/BAM/CRAM files. #
+###############################################
 
-Input files must be specified with either '--bam' or '--fn'.
+Alignment files must be specified with either '--bam' or '--fn', but it's an error to use both.
 
-By default, the command will count all reads within the alignment file. Use '--bed' or '--region' to specify target regions. When you do this, pay attention to the 'chr' string in contig names (e.g. 'chr1' vs. '1').
+By default, the command will count all reads within the alignment files. You can specify target regions with either '--bed' or '--region', but not both. When you do this, pay close attention to the 'chr' string in contig names (e.g. 'chr1' vs. '1'). Note also that '--region' requires the input files be indexed.
+
+Under the hood, the command computes read depth using the 'samtools depth' command.
 
 Usage examples:
   $ fuc {api.common._script_name()} --bam 1.bam 2.bam --bed in.bed > out.tsv
@@ -27,12 +31,12 @@ def create_parser(subparsers):
         '--bam',
         metavar='PATH',
         nargs='+',
-        help='One or more input files.'
+        help='One or more alignment files.'
     )
     parser.add_argument(
         '--fn',
         metavar='PATH',
-        help='File containing one input filename per line.'
+        help='File containing one alignment file per line.'
     )
     parser.add_argument(
         '--bed',
@@ -42,7 +46,7 @@ def create_parser(subparsers):
     parser.add_argument(
         '--region',
         metavar='TEXT',
-        help="Only report depth in specified region ('chrom:start-end')."
+        help="Target region ('chrom:start-end')."
     )
     parser.add_argument(
         '--zero',